+++ sdk/search/azure-search-documents/azure/search/documents/_internal/aio/_search_indexing_buffered_sender_async.py
@@ -34,10 +34,8 @@ class SearchIndexingBufferedSender(SearchIndexingBufferedSenderBase, HeadersMixi
     :type index_name: str
     :param credential: A credential to authorize search client requests
     :type credential: ~azure.core.credentials.AzureKeyCredential
-    :keyword bool auto_flush: if the auto flush mode is on. Default to True.
     :keyword int auto_flush_interval: how many max seconds if between 2 flushes. This only takes effect
-        when auto_flush is on. Default to 60 seconds. If a non-positive number is set, it will be default
-        to 86400s (1 day)
+        when auto_flush is on. Default to 60 seconds.
     :keyword int initial_batch_action_count: The initial number of actions to group into a batch when
         tuning the behavior of the sender. The default value is 512.
     :keyword int max_retries_per_action: The number of times to retry a failed document. The default value is 3.
@@ -106,6 +104,7 @@ class SearchIndexingBufferedSender(SearchIndexingBufferedSenderBase, HeadersMixi
         :param int timeout: time out setting. Default is 86400s (one day)
         :return: True if there are errors. Else False
         :rtype: bool
+        :raises ~azure.core.exceptions.ServiceResponseTimeoutError:
         """
         has_error = False
         begin_time = int(time.time())
@@ -113,6 +112,10 @@ class SearchIndexingBufferedSender(SearchIndexingBufferedSenderBase, HeadersMixi
             now = int(time.time())
             remaining = timeout - (now - begin_time)
             if remaining < 0:
+                if self._on_error:
+                    actions = await self._index_documents_batch.dequeue_actions()
+                    for action in actions:
+                        await self._on_error(action)
                 raise ServiceResponseTimeoutError("Service response time out")
             result = await self._process(timeout=remaining, raise_error=False)
             if result: