+++ sdk/search/azure-search-documents/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 11.1.0 (2021-02-09)
+## 11.1.0 (2021-02-10)
 
 **Breaking Changes**